renderMovieListをlibディレクトリからcomponentsディレクトリに移動、CSS修正
renderMovieListを関数で定義しておりTailwindCSSが効かな異状態だったので、UserRegisteredMovieというコンポーネントファイルに変更
