+++ src/components/UserTabPanel.tsx
@@ -1,6 +1,6 @@
-import { renderMovieList } from '@/lib/userTab/renderMovieList'
 import { Lists } from '@/types/Lists'
 
+import UserRegisteredMovie from './UserRegisteredMovie'
 import UserTabPanelItem from './UserTabPanelItem'
 
 type Props = {
@@ -18,13 +18,16 @@ const UserTabPanel = ({
   return (
     <div className="mt-4 sm:mt-0">
       <UserTabPanelItem tabId="favorites" selectedTab={selectedTab}>
-        {renderMovieList(Object.keys(favorites), getRegisteredMovieData)}
+        {UserRegisteredMovie(Object.keys(favorites), getRegisteredMovieData)}
       </UserTabPanelItem>
       <UserTabPanelItem tabId="watchlists" selectedTab={selectedTab}>
-        {renderMovieList(Object.keys(watchlists), getRegisteredMovieData)}
+        {UserRegisteredMovie(Object.keys(watchlists), getRegisteredMovieData)}
       </UserTabPanelItem>
       <UserTabPanelItem tabId="watched" selectedTab={selectedTab}>
-        {renderMovieList(Object.keys(movieListData), (id) => movieListData[id])}
+        {UserRegisteredMovie(
+          Object.keys(movieListData),
+          (id) => movieListData[id],
+        )}
       </UserTabPanelItem>
     </div>
   )
